python-service/requirements.txt
flask==3.0.3
flask-cors==4.0.1
face_recognition==1.3.0
numpy==1.26.4
pymongo==4.8.0
pillow==10.4.0
python-dotenv==1.0.1
scipy==1.13.1
gunicorn==22.0.0
cmake==3.29.6
dlib==19.24.4